Correct model saving technique flow
What was the problem ?
-----------------------
-l option is used to load an existing pkl file however such a file was
not saved during the training therefore running the play program with -l
won't work.

What did I do ?
----------------

Now a the pickle files are being saved with their respective names under
the training directory. Now -l will be able to load the same during the
play.